requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.8.0
eventlet==0.41.0
fastapi==0.136.0
filelock==3.29.0
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
gevent==26.4.0
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
idna==3.12
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.33
mpmath==1.3.0
multidict==6.7.1
networkx==3.1
numpy==1.26.4
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
openvino==2024.6.0
openvino-dev==2024.6.0
openvino-telemetry==2025.2.0
packaging==26.1
pillow==12.2.0
polars==1.40.0
polars-runtime-32==1.40.0
propcache==0.4.1
psutil==7.2.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
PyYAML==6.0.3
requests==2.33.1
scipy==1.17.1
simple-websocket==1.1.0
six==1.17.0
sounddevice==0.5.5
starlette==1.0.0
sympy==1.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.45.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.8
wsproto==1.3.2
yarl==1.23.0
zope.event==6.1
zope.interface==8.3
